resolve clippy warnings + shift reused lib functions to core
