+++ crates/imgdd/benches/rust_benches.rs
@@ -2,22 +2,6 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use imgdd::*;
 use std::path::PathBuf;
 
-fn benchmark_select_filter_type(c: &mut Criterion) {
-    c.bench_function("select_filter_type", |b| {
-        b.iter(|| {
-            black_box(select_filter_type(Some("nearest")));
-        });
-    });
-}
-
-fn benchmark_select_algo(c: &mut Criterion) {
-    c.bench_function("select_algo", |b| {
-        b.iter(|| {
-            black_box(select_algo(Some("dhash")));
-        });
-    });
-}
-
 fn benchmark_hash(c: &mut Criterion) {
     let dir_path = PathBuf::from("../../imgs/test/single");
 
@@ -52,8 +36,6 @@ fn benchmark_dupes(c: &mut Criterion) {
 
 criterion_group!(
     rust_interface_benchmarks,
-    benchmark_select_filter_type,
-    benchmark_select_algo,
     benchmark_hash,
     benchmark_dupes
 );
