+++ crates/imgdd/src/lib.rs
@@ -2,55 +2,11 @@
 //! Leverages perceptual hashing algorithms to identify duplicate or visually similar images in a directory.
 
 use anyhow::Error;
-use image::imageops::FilterType;
-use imgddcore::dedupe::*;
-use imgddcore::validate::*;
+use imgddcore::dedupe::{collect_hashes, find_duplicates, sort_hashes};
+use imgddcore::utils::{select_algo, select_filter_type};
+use imgddcore::validate::validate_path;
 use std::collections::HashMap;
 use std::path::PathBuf;
-
-/// Converts a string to a `FilterType`.
-///
-/// # Arguments
-///
-/// - `filter` - String specifying the filter type.
-///     - **Options:** [`Nearest`, `Triangle`, `CatmullRom`, `Gaussian`, `Lanczos3`]
-///
-/// # Returns
-///
-/// - A `FilterType` enum corresponding to the input string.
-#[inline]
-pub fn select_filter_type(filter: Option<&str>) -> FilterType {
-    match filter.unwrap_or("nearest") {
-        ref f if f.eq_ignore_ascii_case("nearest") => FilterType::Nearest,
-        ref f if f.eq_ignore_ascii_case("triangle") => FilterType::Triangle,
-        ref f if f.eq_ignore_ascii_case("catmullrom") => FilterType::CatmullRom,
-        ref f if f.eq_ignore_ascii_case("gaussian") => FilterType::Gaussian,
-        ref f if f.eq_ignore_ascii_case("lanczos3") => FilterType::Lanczos3,
-        other => panic!("Unsupported filter type: {}", other),
-    }
-}
-
-/// Selects a hashing algorithm.
-///
-/// # Arguments
-///
-/// - `algo` - String specifying the hashing algorithm.
-///     - **Options:** [`aHash`, `mHash`, `dHash`, `pHash`, `wHash`]
-///
-/// # Returns
-///
-/// - A standardized `&'static str` representing the selected algorithm.
-#[inline]
-pub fn select_algo(algo: Option<&str>) -> &'static str {
-    match algo.unwrap_or("dhash") {
-        input if input.eq_ignore_ascii_case("dhash") => "dhash",
-        input if input.eq_ignore_ascii_case("ahash") => "ahash",
-        input if input.eq_ignore_ascii_case("mhash") => "mhash",
-        input if input.eq_ignore_ascii_case("phash") => "phash",
-        input if input.eq_ignore_ascii_case("whash") => "whash",
-        other => panic!("Unsupported algorithm: {}", other),
-    }
-}
 
 /// Calculates hashes for all images in a directory recursively.
 ///
@@ -148,5 +104,6 @@ pub fn dupes(
     let mut hash_paths = collect_hashes(validated_path, filter_type, selected_algo)?;
     sort_hashes(&mut hash_paths);
 
-    Ok(find_duplicates(&hash_paths, remove)?)
+    // Ok(find_duplicates(&hash_paths, remove)?)
+    find_duplicates(&hash_paths, remove)
 }
