+++ crates/imgddcore/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod dedupe;
 pub mod hashing;
 pub mod normalize;
+pub mod utils;
 pub mod validate;
