+++ crates/imgddpy/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "imgddpy"
-version = "0.1.5"
+version = "0.1.6"
 edition.workspace = true
 license.workspace = true
 authors.workspace = true
@@ -12,7 +12,7 @@ homepage.workspace = true
 readme = "README.md"
 
 [dependencies]
-imgddcore = { path = "../imgddcore", version = "0.1.3" }
+imgddcore = { path = "../imgddcore", version = "0.1.4" }
 pyo3 = { version = "0.23", features = ["extension-module", "abi3-py39"] }
 image.workspace = true
