+++ crates/imgddpy/src/lib.rs
@@ -1,33 +1,9 @@
-use image::imageops::FilterType;
-use imgddcore::dedupe::*;
-use imgddcore::validate::*;
+use imgddcore::dedupe::{collect_hashes, find_duplicates, sort_hashes};
+use imgddcore::utils::{select_algo, select_filter_type};
+use imgddcore::validate::validate_path;
 use pyo3::prelude::*;
 use std::collections::HashMap;
 use std::path::PathBuf;
-
-#[inline]
-fn select_filter_type(filter: Option<&str>) -> FilterType {
-    match filter.unwrap_or("nearest") {
-        ref f if f.eq_ignore_ascii_case("nearest") => FilterType::Nearest,
-        ref f if f.eq_ignore_ascii_case("triangle") => FilterType::Triangle,
-        ref f if f.eq_ignore_ascii_case("catmullrom") => FilterType::CatmullRom,
-        ref f if f.eq_ignore_ascii_case("gaussian") => FilterType::Gaussian,
-        ref f if f.eq_ignore_ascii_case("lanczos3") => FilterType::Lanczos3,
-        other => panic!("Unsupported filter type: {}", other),
-    }
-}
-
-#[inline]
-fn select_algo(algo: Option<&str>) -> &'static str {
-    match algo.unwrap_or("dhash") {
-        input if input.eq_ignore_ascii_case("dhash") => "dhash",
-        input if input.eq_ignore_ascii_case("ahash") => "ahash",
-        input if input.eq_ignore_ascii_case("mhash") => "mhash",
-        input if input.eq_ignore_ascii_case("phash") => "phash",
-        input if input.eq_ignore_ascii_case("whash") => "whash",
-        other => panic!("Unsupported algorithm: {}", other),
-    }
-}
 
 /// ```python
 /// hash(path, filter="triangle", algo="dhash", sort=False)
@@ -73,7 +49,7 @@ pub fn hash(
     let filter_type = select_filter_type(filter);
     let algo = select_algo(algo);
 
-    let mut hash_paths = collect_hashes(&validated_path, filter_type, &algo)
+    let mut hash_paths = collect_hashes(validated_path, filter_type, algo)
         .map_err(|e| PyErr::new::<pyo3::exceptions::PyRuntimeError, _>(format!("{}", e)))?;
 
     // Optionally sort hashes
@@ -131,7 +107,7 @@ pub fn dupes(
     let filter_type = select_filter_type(filter);
     let algo = select_algo(algo);
 
-    let mut hash_paths = collect_hashes(&validated_path, filter_type, &algo)
+    let mut hash_paths = collect_hashes(validated_path, filter_type, algo)
         .map_err(|e| PyErr::new::<pyo3::exceptions::PyRuntimeError, _>(format!("{}", e)))?;
     sort_hashes(&mut hash_paths);
